+++ src/test/java/io/github/lilaschuda/guanaco/core/GuanacoRouteBuilderTestSupport.java
@@ -38,7 +38,7 @@ public abstract class GuanacoRouteBuilderTestSupport {
             RouteConfig config,
             String processorName,
             RouteOutcomeRegistry registry) throws Exception {
-        registerRoute(processor, routeInterface, config, processorName, registry, Map.of());
+        registerRoute(processor, routeInterface, config, processorName, registry, Map.of(), Map.of());
     }
 
     protected void registerRoute(
@@ -47,9 +47,10 @@ public abstract class GuanacoRouteBuilderTestSupport {
             RouteConfig config,
             String processorName,
             RouteOutcomeRegistry registry,
-            Map<String, AggregationStrategy> aggregationStrategies) throws Exception {
+            Map<String, AggregationStrategy> aggregationStrategies,
+            Map<String, GuanacoDelayStrategy> delayStrategies) throws Exception {
         context.addRoutes(new GuanacoRouteBuilder(
-                processor, routeInterface, config, processorName, registry, aggregationStrategies));
+                processor, routeInterface, config, processorName, registry, aggregationStrategies, delayStrategies));
     }
 
     protected void registerRoute(
@@ -57,7 +58,7 @@ public abstract class GuanacoRouteBuilderTestSupport {
             RouteConfig config,
             String processorName,
             RouteOutcomeRegistry registry) throws Exception {
-        registerRoute(processor, ROUTE_OUTCOME_CLASS, config, processorName, registry, Map.of());
+        registerRoute(processor, ROUTE_OUTCOME_CLASS, config, processorName, registry, Map.of(), Map.of());
     }
 
     /**